% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndviMonitor-package.R
\docType{package}
\name{ndviMonitor-package}
\alias{ndviMonitor}
\alias{ndviMonitor-package}
\title{ndviMonitor : Suivi phénologique des cultures par séries temporelles NDVI}
\description{
\code{ndviMonitor} propose un flux de travail simple et reproductible pour
suivre la phénologie des cultures à partir d'images Sentinel-2 :
import de parcelles agricoles, téléchargement d'images Sentinel-2,
calcul du NDVI, extraction de statistiques zonales par parcelle,
lissage des séries temporelles, détection de stades phénologiques
simples et d'anomalies de stress, visualisations et rapport HTML
automatique.
}
\section{Fonctions principales}{

\describe{
\item{\code{\link[=import_parcels]{import_parcels()}}}{Importer des parcelles (shapefile/GeoJSON).}
\item{\code{\link[=download_sentinel]{download_sentinel()}}}{Télécharger des images Sentinel-2
(Microsoft Planetary Computer).}
\item{\code{\link[=calc_ndvi]{calc_ndvi()}}}{Calculer le NDVI à partir des bandes Rouge/NIR.}
\item{\code{\link[=extract_ndvi]{extract_ndvi()}}}{Extraire les statistiques NDVI par parcelle.}
\item{\code{\link[=smooth_timeseries]{smooth_timeseries()}}}{Lisser les séries temporelles NDVI.}
\item{\code{\link[=detect_growth_stage]{detect_growth_stage()}}}{Détecter les stades phénologiques.}
\item{\code{\link[=detect_stress]{detect_stress()}}}{Détecter des anomalies de stress.}
\item{\code{\link[=plot_ndvi_curve]{plot_ndvi_curve()}}}{Tracer les courbes NDVI.}
\item{\code{\link[=plot_ndvi_map]{plot_ndvi_map()}}}{Cartographier le NDVI.}
\item{\code{\link[=generate_report]{generate_report()}}}{Générer un rapport HTML automatique.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/votre-utilisateur/ndviMonitor}
  \item Report bugs at \url{https://github.com/votre-utilisateur/ndviMonitor/issues}
}

}
\author{
\strong{Maintainer}: Etudiant Projet \email{etudiant@example.com}

}
\keyword{internal}
